Add a garbage collector
Doesn't work tho (the refcell test fails currently).

While I'm busy fixing whatever's wrong with this code, here's the
general overview. So it's like an Rc/RefCell, with reference counts and
all, but then there's also mark-and-sweep that solves the reference
cycle problem. Additionally, like any good GC, it has finalization
(full of footguns, obviously). It should even handle resurrection!

...Probably. I don't see why not, but then again, there's a dozen or so
of `unsafe` blocks in the code, each of them eager to prove me wrong.
Reminds me of writing C, except I don't think it was this stressful.
(Rust does its job deterring people from the unsafe a tad too well imo.)

Anyway, it also turned out to be quite similar to the gc crate.
This was not intentional. I guess this kind of validates the design?
Maybe I'll just switch to that crate later. We'll see.