Add a new (old) crate
While I'm ironing out the design, it's a great time to do something that
requires no input from one's overworked brain. Like this commit, for
example! This crate I'm christening stitch because something has to be
called that or the repository name wouldn't make any sense, and it was
quite literally a no-brainer for that reason.