Wasm IR, take two
Decided to scratch all the code I've been working on for the past week
and start anew. This time it's going to be an AST-based IR, and I'm
defining it all myself instead of piggybacking on third-party crates.
With any luck I'll manage to produce something usable (not necessarily
useful) by the end of the year to evaluate just how bad it will perform.