Perform symbolic memory loads and stores
While it does eliminate a few load/store instructions from the output,
the primary goal pursued by this work has not been achieved: calc.wat's
value stack (used in $interpret to hold intermediate values) operations
are still present in the result.

The problem is that the stack's backing storage is allocated via $malloc
and a pointer to the allocation is then written to memory, whereupon it
loses the metadata that allows the allocation to be tracked
symbolically. When $vec-push loads the pointer, it gets a plain I32 that
it can't safely dereference, thus requiring emitting loads and stores
targetting it.

On the other hand, the two other fields of the stack — its length and
capacity — should have been possible to track during specialization. The
fact that the output still contains rolled-up $memset loops points
otherwise and it implies that the values of these fields were not after
all available to the specializer (as the loop exit conditionals could
not be eliminated). This behavior is unexpected and needs investigation.