Make stronger assumptions for stitch/symbolic-ptr
Now we additionally assume no one has an access to an allocation pointed
to by a symbolic pointer created with stitch/symbolic-ptr. This
assumption can be easily broken if the symbolic pointer is written to a
global or memory, so it's rather dangerous. However, now the specialized
calc.wat is almost linear! The only branches are due to memory.size
checks.